backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
pandas==2.1.3
python-dateutil==2.8.2
pytest==7.4.3
httpx==0.25.2
sendgrid==6.11.0
